defconfig: Disable cpu boost
This is totally broken by oplus's ormsHalService. It set 3GHz boost freq
for all cores, and that of small cluster was then override by post_boot,
still leaving prime and big cluster on a unreasonable boost freq. This
makes big and prime cluster constantly boost to max freq on input event,
which could consume a lot of extra power. Disable it to resolve this
situation from the root cause.

[   13.490699] cpu-boost: vendor.oplus.ha(1222) is setting 1324800 for cpu0
[   13.490700] cpu-boost: vendor.oplus.ha(1222) is setting 0 for cpu4
[   13.490702] cpu-boost: vendor.oplus.ha(1222) is setting 0 for cpu7
[   14.140920] cpu-boost: vendor.oplus.ha(1222) is setting 3000000 for cpu0
[   14.140923] cpu-boost: vendor.oplus.ha(1222) is setting 3000000 for cpu4
[   14.140924] cpu-boost: vendor.oplus.ha(1222) is setting 3000000 for cpu7
[   17.358046] cpu-boost: init.qcom.post_(4740) is setting 1324800 for cpu0

Signed-off-by: LibXZR <i@xzr.moe>